pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.8",
    "fastapi[standard]>=0.124.0",
    "google-genai>=1.54.0",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0",
    "uvicorn>=0.38.0",
]
